+++ src/go/rpk/pkg/adminapi/admin.go
@@ -23,7 +23,7 @@ import (
 	mTerm "github.com/moby/term"
 	"github.com/redpanda-data/common-go/rpadmin"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth"
+	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth/authtoken"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth/providers/auth0"
 	"github.com/spf13/afero"
 	"go.uber.org/zap"
@@ -52,13 +52,13 @@ func GetAuth(p *config.RpkProfile) (rpadmin.Auth, error) {
 		if a == nil || len(a.AuthToken) == 0 {
 			return nil, errors.New("no current auth found, please login with 'rpk cloud login'")
 		}
-		expired, err := oauth.ValidateToken(
+		expired, err := authtoken.ValidateToken(
 			a.AuthToken,
 			auth0.NewClient(p.DevOverrides()).Audience(),
 			a.ClientID,
 		)
 		if err != nil {
-			if errors.Is(err, oauth.ErrMissingToken) {
+			if errors.Is(err, authtoken.ErrMissingToken) {
 				return nil, err
 			}
 			return nil, fmt.Errorf("unable to validate cloud token, please login again using 'rpk cloud login': %v", err)
