+++ src/go/rpk/pkg/cli/cloud/BUILD
@@ -19,6 +19,7 @@ go_library(
         "//src/go/rpk/pkg/cli/profile",
         "//src/go/rpk/pkg/config",
         "//src/go/rpk/pkg/oauth",
+        "//src/go/rpk/pkg/oauth/authtoken",
         "//src/go/rpk/pkg/oauth/providers/auth0",
         "//src/go/rpk/pkg/os",
         "//src/go/rpk/pkg/out",
