+++ src/go/rpk/pkg/cli/profile/BUILD
@@ -26,7 +26,7 @@ go_library(
         "//src/go/rpk/pkg/cli/container/common",
         "//src/go/rpk/pkg/cobraext",
         "//src/go/rpk/pkg/config",
-        "//src/go/rpk/pkg/oauth",
+        "//src/go/rpk/pkg/oauth/authtoken",
         "//src/go/rpk/pkg/oauth/providers/auth0",
         "//src/go/rpk/pkg/os",
         "//src/go/rpk/pkg/out",
