+++ src/go/rpk/pkg/oauth/BUILD
@@ -11,11 +11,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/go/rpk/pkg/config",
+        "//src/go/rpk/pkg/oauth/authtoken",
         "//src/go/rpk/pkg/os",
         "//src/go/rpk/pkg/publicapi",
         "@build_buf_gen_go_redpandadata_cloud_protocolbuffers_go//redpanda/api/iam/v1:iam",
         "@com_connectrpc_connect//:connect",
-        "@com_github_lestrrat_go_jwx_v2//jwt",
         "@com_github_spf13_afero//:afero",
         "@org_uber_go_zap//:zap",
     ],
@@ -32,8 +32,6 @@ go_test(
     deps = [
         "//src/go/rpk/pkg/config",
         "@build_buf_gen_go_redpandadata_cloud_protocolbuffers_go//redpanda/api/iam/v1:iam",
-        "@com_github_lestrrat_go_jwx_v2//jwa",
-        "@com_github_lestrrat_go_jwx_v2//jwt",
         "@com_github_spf13_afero//:afero",
         "@com_github_stretchr_testify//require",
         "@org_golang_google_protobuf//proto",
