+++ src/go/rpk/pkg/schemaregistry/BUILD
@@ -9,7 +9,7 @@ go_library(
         "//src/go/rpk/pkg/adminapi",
         "//src/go/rpk/pkg/config",
         "//src/go/rpk/pkg/net",
-        "//src/go/rpk/pkg/oauth",
+        "//src/go/rpk/pkg/oauth/authtoken",
         "//src/go/rpk/pkg/oauth/providers/auth0",
         "@com_github_redpanda_data_common_go_rpsr//:rpsr",
         "@com_github_spf13_afero//:afero",
