+++ src/go/rpk/pkg/schemaregistry/client.go
@@ -9,7 +9,7 @@ import (
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/adminapi"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/net"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth"
+	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth/authtoken"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/oauth/providers/auth0"
 	"github.com/spf13/afero"
 	"github.com/twmb/franz-go/pkg/kgo"
@@ -68,13 +68,13 @@ func NewClient(fs afero.Fs, p *config.RpkProfile) (*rpsr.Client, error) {
 		if a == nil || len(a.AuthToken) == 0 {
 			return nil, errors.New("no current cloud token found in your profile, please login with 'rpk cloud login'")
 		}
-		expired, err := oauth.ValidateToken(
+		expired, err := authtoken.ValidateToken(
 			a.AuthToken,
 			auth0.NewClient(p.DevOverrides()).Audience(),
 			a.ClientID,
 		)
 		if err != nil {
-			if errors.Is(err, oauth.ErrMissingToken) {
+			if errors.Is(err, authtoken.ErrMissingToken) {
 				return nil, err
 			}
 			return nil, fmt.Errorf("unable to validate cloud token, please login again using 'rpk cloud login': %v", err)
